+++ test/tint/buffer/uniform/std140/unnested/mat2x2/dynamic_index_via_ptr.wgsl.expected.glsl
@@ -35,9 +35,9 @@ vec2 load_m_inner_p0(uint p0) {
 void f() {
   mat2 p_m = load_m_inner();
   int tint_symbol = i();
-  vec2 p_m_i = load_m_inner_p0(uint(tint_symbol));
+  vec2 p_m_i = load_m_inner_p0(uint(uint(tint_symbol)));
   mat2 l_m = load_m_inner();
-  vec2 l_m_i = load_m_inner_p0(uint(tint_symbol));
+  vec2 l_m_i = load_m_inner_p0(uint(uint(tint_symbol)));
 }
 
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
